spl: fit: Allow the board to tell if more images must be loaded from FIT
spl_fit_get_image_name() is used to get the names of the images that the
SPL must load from the FIT. It relies on the content of a property present
in the FIT. The list of images is thus statically defined in the FIT.
With this scheme, it quickly becomes hard to manage combinations of more
than a handful of images.
To address this problem, give the board driver code the opportunity to
add to the list of images. The images from the FIT property are loaded
first, and then the board_get_fit_loadable() is called to get more image
names.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>

Series-changes:6
- Instead of matching a overlay by the name of it's node. Try to match it
  first with the description, and then with the node's name. This allows
  to use makeimg to add the overlays to u-boot.img and not use a custom
  SPL_FIT_GENERATOR script

Series-changes:5
-reword commit log

Series-to: sjg@chromium.org, trini@konsulko.com, michal.simek@xilinx.com
Series-cc: lokeshvutla@ti.com
Series-version: 6
Series-prefix: PATCH
Series-name: overlays in SPL
Cover-letter:
Add support for applications of overlays in SPL

The purpose of this series is to provide the SPL with ability to apply
overlays for u-boot.

Our use-case is the support of the daughter boards of the AM65x and J721e
EVMs. In Linux, each board is supported by a unique overlay. The presence
of the boards is detected at runtime, and some useful features (like USB)
are implemented on those daughter boards. Instead of providing multiple
dtbs and fall in a combinatorial pit, we propose to use DT overlays.

On arm, if overlay are supported, this series increases the size of the SPL
by 3.2 kB.

Travis build : https://travis-ci.org/jjhiblot/u-boot/builds/600742163
END

Series-changes: 4
- Use the board driver infrastructure to get the image names from the
board code.
- Remove a patch that passed the board name to the FIT generator. If needed
the generator can get it from elsewhere
- Add a fix to not load the firmware twice (once as a firmware and once as
a loadable)

Series-changes:3
- removed the RFC prefix. This work will be needed soon by TI's AM65x
platform. and can probably benefit other modular platforms
- removed the last patch that provided an example of how to use this with
on a DRA76.
- removed the patch that made u-boot.img a symlink to u-boot.itb because
it breaks the build of many platforms (because files required to build the
ITB are missing)
- removed the patch to reduce the footprint of the am335x SPL. (already
merged)
- Made the boot flow more permissive (don't fail immediately if an overlay
is not present) and more verbose when an error occures
- handle the dependencies of the FIT generation in a more generic way
- use a dedicated kconfig option to enable the application of the overlays
by the SPL.

Series-changes:2
- reworked board_fit_get_additionnal_images() and how it used in spl_fit.c
- removed dtbo generation from dtso files and use .dts extension for the
  overlays
- add dynamic allocation usage in a separate patch
- defconfig change for the am335x_evm