sp: fit: Allow the board to dynamically select the DTB overlays it needs.
Currently the list of DTB overlays to apply on top of the DTB is described
in a configuration node. With this scheme, it becomes quickly hard to
manage combinations of more than a hand few of DTB overlays. Instead the
board could tell for each overlay if it is needed or not.
This is the role of board_fit_get_additionnal_images().
Note that it is not limited to dtb overlays, it could be used for kind of
images (fpga, loadables, etc.)

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>