spl: fit: allocate a temporary buffer to load the overlays
If the node describing an overlay does not specify a load address, it will
be loaded at the address previously used.
Fixing it by allocating a temporary buffer that will be used as a
default load address. By default, the size of the buffer is 64kB which
should be plenty for most use cases.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>

Series-changes:5
- Do not allocate the buffer if not needed (no overlay).
- Add a Kconfig option for the buffer size

Series-changes:4
- make sure that the temp buffer is freed in all cases