ARM: mach-k3: Provide a FIT generator script
The FIT generator script is used to generated a ITS file that is, in turn,
used to produce a FIT image more complex than the default provided by
mkimage.
This generator produces a FIT image containing u-boot, the dtbs
in OF_LIST and all the DT overlays used by the board.
The list of overlays is provided by CONFIG_OF_OVERLAY_LIST

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>