test: clk: test clock self assignment
Make sure that the clock self-assignment works by having a clock of
clk-sbox be configured automatically when clk-sbox is probed.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>

Series-to: sjg@chromium.org, trini@konsulko.com
Series-cc: lokeshvutla@ti.com
Series-version: 1
Series-name: clk_changes
Cover-letter:
clk: Add a managed API and fix clock self-assignment

This is the second of a few series, the goal of which is to facilitate
porting drivers from the linux kernel. Most of the series will be about
adding managed API to existing infrastructure (GPIO, reset, phy,...)

This particular series is about clocks. It adds a managed API, using
the same API as linux. It also fix a use-case where the clock is
initialized (rate and/or mux) when it is probed using the
assigned-clocks-* properties.

END