test: reset: Add tests for the managed API
The tests are basically the same as for the regular API. Except that
the reset are initialized using the managed API, and no freed manually.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>

Series-to: sjg@chromium.org, trini@konsulko.com
Series-cc: lokeshvutla@ti.com
Series-version: 1
Series-name: reset_changes
Cover-letter:
reset: Add a managed API

This is the 3rd of a few series, the goal of which is to facilitate
porting drivers from the linux kernel. Most of the series will be about
adding managed API to existing infrastructure (GPIO, reset, phy,...)

This particular series is about reset controllers. It adds a managed API,
close to that of linux. The main difference is that bulk and reset_ctl
are handled with different functions.
END