test: Add tests for the multiplexer framework
Provide tests to check the behavior of the multiplexer framework.
The test uses a mmio-based multiplexer.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>

Series-to: sjg@chromium.org, trini@konsulko.com
Series-cc: lokeshvutla@ti.com
Series-version: 1
Series-name: gpio_changes
Cover-letter:
drivers: Add a framework for MUX drivers

Add a new minimalistic subsystem that handles multiplexer controllers.
It provides the same API as Linux and mux drivers should be portable with
a minimum effort.
This series also includes a port of the Linux's mmio-mux driver.

This series relies on 2 other series [1] and [2] posted earlier.

[1] : https://patchwork.ozlabs.org/project/uboot/list/?series=132923
[2] : https://patchwork.ozlabs.org/project/uboot/list/?series=133713
END